requirements.txt
jupyterlab==3.5.2
matplotlib==3.6.3
numba==0.56.4
numpy==1.23.5
pandas==1.5.2
Pillow==9.4.0
pytorch-lightning==1.8.6
scikit-learn==1.2.0
scipy==1.10.0
tensorboardX==2.5.1
timm==0.6.12
torch==1.13.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.9
plotly==5.14.1
kaleido==0.2.1
